refactor(worker): Split task executors
